check_op_code.sh: Add OpName to lists that are checked for alphabetic order. op_code.hpp: Alphabatize OpName.
git-svn-id: https://projects.coin-or.org/svn/CppAD/trunk@2027 0e562018-b8fb-0310-a933-ecf073c7c197

+++ bin/check_op_code.sh
@@ -55,6 +55,17 @@ then
 	exit 1
 fi
 # -----------------------------------------------------------------------------
+# check OpName 
+sed -n -e '/char \*OpName\[\]/,/^[ 	]*};/p' cppad/local/op_code.hpp | \
+	sed -e '/char \*OpName\[\]/d' -e '/^[ 	]*};/d' \
+	-e 's|^[ 	]*"||' -e 's|".*|Op|' > bin/op_code.5.$$
+#
+if ! diff bin/op_code.1.$$ bin/op_code.5.$$ 
+then
+	echo "check_op_code.sh: OpName list is not in alphabetical order"
+	exit 1
+fi
+# -----------------------------------------------------------------------------
 # clean up
 rm bin/op_code.*.$$
 echo "bin/check_op_code.sh: OK"
